backend/requirements.txt
fastapi==0.131.0
uvicorn==0.41.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.1
python-multipart==0.0.22
bcrypt==4.3.0
